requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
bauh==0.10.5
beautifulsoup4==4.11.2
bleach==6.0.0
btrfsutil==6.2.2
buildozer==1.5.1.dev0
cachetools==5.3.0
cairocffi==1.4.0
CairoSVG==2.5.2
catboost==1.1.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
cmdstanpy==1.1.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.3
configobj==5.1.0.dev0
constantly==15.1.0
contourpy==1.0.7
convertdate==2.4.0
cryptography==40.0.1
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.34
darts==0.23.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deluge==2.1.1
deprecation==2.1.0
distlib==0.3.6
distro==1.8.0
docutils==0.19
entrypoints==0.4
ephem==4.1.4
evdev==1.6.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.1.0
GeoIP==1.3.2
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-speech==2.19.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
gTTS==2.3.1
guiscrcpy==2023.1.1
hijri-converter==2.2.4
holidays==0.18
httplib2==0.22.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
ifaddr==0.2.0
IMDbPY==2021.4.18
incremental==22.10.0
inflect==6.0.4
injector==0.20.1
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.13
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_packaging==0.12.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lensfun==0.3.3
libinput-gestures-qt==0.4
libtorrent==2.0.8
lightgbm==3.3.5
lightning-utilities==0.6.0.post0
llvmlite==0.39.1
LunarCalendar==0.0.9
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mercurial==6.4.2
mistune==2.0.5
more-itertools==9.1.0
msgpack==1.0.4
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
netsnmp-python==1.0a1
nfoursid==1.0.1
nftables==0.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.27.4
openshot-qt==3.0.0
OpenTimelineIO==0.15.0
optuna==3.1.0
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peewee==3.16.1
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
playsound==1.3.0
plotly==5.13.0
ply==3.11
pmdarima==2.0.2
pocketsphinx==5.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
prophet==1.1.2
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pvporcupine==2.2.0
pvrecorder==1.1.1
py3nvml==0.2.7
pyaaf2==1.6.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycairo==1.23.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.44.1
pyinotify==0.9.6
PyMeeus==0.5.12
pynput==1.7.6
pyod==1.0.7
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
pyrsistent==0.19.3
PySide2==5.15.9
pysmbc==1.0.23
python-dateutil==2.8.2
python-distutils-extra==2.39
python-json-logger==2.0.7
python-pam==2.0.2
python-xlib==0.33
pytorch-lightning==1.9.0
pyttsx3==2.90
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
pyzmq==25.0.2
QTermWidget==1.3.0
QtPy==2.3.1
ray==2.2.0
Reflector==2021.11.20.2.41.3
rencode==1.0.6
requests==2.28.2
requests-file==1.5.1
requests-unixsocket==0.3.0
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Rx==3.2.0
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.0
service-identity==21.1.0
setproctitle==1.3.2
sh==1.14.3
shap==0.41.0
shiboken2==5.15.9
six==1.16.0
slicer==0.0.7
sniffio==1.3.0
soundconverter==4.0.3
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4
SpeechRecognition==3.10.0
SQLAlchemy==1.4.44
stack-data==0.6.2
statsforecast==1.4.0
statsmodels==0.13.5
tabulate==0.9.0
tbats==1.1.2
TBB==0.2
team==1.0
tenacity==8.1.0
tensorboardX==2.5.1
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
torch==1.13.1
torchmetrics==0.11.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trove-classifiers==2023.4.20
Twisted==22.10.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
ufw==0.36.1
Unidecode==1.3.6
urllib3==1.26.13
validate==5.1.0.dev0
validate-pyproject==0.12.2.post1.dev0+g2940279.d20230328
variety==0.8.9
vboxapi==1.0
virtualenv==20.17.1
wavio==0.0.7
wcwidth==0.2.5
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.5.1
widgetsnbextension==4.0.7
xarray==2023.1.0
xcffib==1.2.0
xgboost==1.7.3
yarl==1.8.2
youtube-dl==2021.12.17
zope.interface==6.0
